Install colcon-bash when building rolling-on-focal (#205)
One of the goals of the CI jobs is to create a minimal environment in an
effort to assert dependencies are declared properly. The purpose of the
rolling-on-focal job is to create a distributable archive for use in the
same way our archives produced by ci.ros2.org are created.

To that end, we should install the not-strictly-necessary
python3-colcon-bash package to ensure that a setup.bash script is put in
the resulting archive, which aligns with the ci.ros2.org archives.