Fix the list_priorities script to work again. (#232)
In particular, we needed to make sure that we support !include
in YAML, which is new since the last time this script was
touched (in 2016!).  Along with that, revamp the whole
structure of it, since we now have to be more careful about
directories.  And finally, make it very specifically look at
particular directories, since nowadays not all directories
contain jenkins config scripts.

With these changes, it can now be used again to list out the
priorities.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>