Standardize the binary job timeouts at 150 minutes. (#311)
Currently, Ubuntu amd64 and RHEL amd64 binary jobs are
at 120 minutes, while Ubuntu aarch64 binary jobs are at 720
minutes.  That latter timeout is probably a legacy of when
the aarch64 jobs used qemu to build.

This controversial PR changes all of them to 150 minutes.
To be clear, that is a 30 minute *increase* in the case of
amd64, and a 570 minute decrease in the case of aarch64.

My reasoning here is as follows:
1.  We know we have jobs on the buildfarm that "sometimes"
complete in their current 120 minute timeslot.  When they
run out of time, they usually would have finished in the
next 10 minutes.
2.  aarch64 jobs have been set to 720 minutes for a while,
but we don't see a huge number of binary jobs taking longer
than 150 minutes there.  So I think our risk of lots of
packages suddenly increasing their time is limited.
3.  This should materially improve the experience of the ROS
Bosses, as there will be fewer jobs they have to run "by hand"
to have regression-free syncs.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>